2010-03-04 11:06 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/include/hbcompdf.h   * harbour/src/vm/macro.c     % changed HB_PCODE_INFO structure to eliminate one memory allocation       in macro compiler
  * harbour/include/hbapi.h
  * harbour/include/hbvm.h
  * harbour/src/vm/codebloc.c
  * harbour/src/vm/itemapi.c
  * harbour/src/vm/hvm.c
    * changed second parameter in hb_codeblockGet*() function to 'int'
    * merged codeblock execution context setting into one function
      hb_vmDoBlock() and removed hb_codeblockEvaluate()
    + added new internal function hb_vmEval()

  * harbour/src/vm/arrays.c
    % use hb_vmEval()

  * harbour/src/macro/macrolex.c
    * added missing const to maro text declaration

  * harbour/tests/speedtst.prg
    ! fixed to work with CLIP and xHarbour after we changed
      SECONDSCPU() to HB_SECONDSCPU() - all these compilers and FlagShip
      have SECONDSCPU() function

+++ harbour/src/vm/itemapi.c
@@ -1845,7 +1845,7 @@ PHB_ITEM hb_itemUnRefOnce( PHB_ITEM pItem )
          {
             /* local variable referenced in a codeblock */
             pItem = hb_codeblockGetRef( pItem->item.asRefer.BasePtr.block,
-                                        pItem->item.asRefer.value );
+                                        ( int ) pItem->item.asRefer.value );
          }
       }
    }
